Changer la structure du backend et ajouter des tests (#8)
* Refactor la structure des fichiers pour Bucket

* Refactor de ce qui est en lien avec "sensors"

* Supprimer le dossier "bucket"

* Ajout de tests en lien avec les sensors

* Remettre les "#" retirés lors du refactor précédent
