+++ serveur-web/api-v2/router/bucket.py
@@ -1,0 +1,109 @@
+# Copyright (C) 2018 Roch D'Amour <roch.damour@gmail.com>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from flask import Blueprint
+from flask import request
+import json
+
+from datetime import datetime, timezone
+
+from controller.bucket import BucketController
+
+# Create a bucket_bluprint, which make it exportable for other modules to use.
+# We bind new routes to this blueprint, which will then be used by the
+# main Flask process when handling requests.
+bucket_blueprint = Blueprint("bucket_blueprint", __name__)
+
+# The following three functions are used to create modification commands
+"""
+GET => /buckets
+Envoie la liste des buckets.
+"""
+@bucket_blueprint.route("/buckets", methods=["GET"])
+def index():
+	return BucketController.index()
+
+
+"""
+POST => /buckets
+Recoit les informations pour enregistrer un nouveau bucket
+"""
+@bucket_blueprint.route("/buckets", methods=["POST"])
+def create():
+	print("Acces a /bucket avec POST")
+	try:
+		if request.headers['Content-Type'] != 'application/json':
+			response = {
+				"error":1,
+				"message":"Content-Type is not application/json"
+				}
+			return (response, 400)
+		elif request.is_json:
+			return BucketController.create(request)
+		else:
+			raise Exception()
+	except Exception as e:
+		 print("ERROR: Request is not JSON or has missing fields.")
+		 print(e)
+		 response = {
+			 "error":1,
+			 "message":"Missing fields in JSON"
+			 }
+		 return (response, 404)
+
+"""
+GET => /buckets/id
+Renvoie un bucket spécifique.
+"""
+@bucket_blueprint.route("/buckets/<id>", methods=["GET"])
+def id(id):
+	return BucketController.get(id)
+
+
+"""
+POST => /buckets/id
+Met a jours les informations d'un bucket
+"""
+@bucket_blueprint.route("/buckets/<id>", methods=["POST"])
+def update(id):
+	print("Acces a /bucket/id avec POST")
+	try:
+		if request.headers['Content-Type'] != 'application/json':
+			response = {
+				"error":1,
+				"message":"Content-Type is not application/json"
+				}
+			return (response, 400)
+		elif request.is_json:
+			return BucketController.update(request)
+		else:
+			raise Exception()
+	except Exception as e:
+		print("ERROR: Request is not JSON or has missing fields.")
+		response = {
+			"error": 1,
+			"message": "Missing fields in JSON"
+			}
+		print(response)
+		return (response, 404)
+
+
+"""
+DELETE => /buckets/id
+Supprime un bucket
+"""
+@bucket_blueprint.route("/buckets/<id>", methods=["DELETE"])
+def delete(id):
+	return BucketController.delete(id)
